ft/adds file transfer (partial)

+++ src/pb.rs
@@ -38,8 +38,8 @@ pub mod get_file_response {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct FileMetaData {
-    #[prost(uint32, tag = "1")]
-    pub file_size: u32,
+    #[prost(uint64, tag = "1")]
+    pub file_size: u64,
     #[prost(string, tag = "2")]
     pub hash: ::prost::alloc::string::String,
 }
